requirements.txt
Django==5.1.1
djangorestframework==3.14.0
pandas==2.1.4; sys_platform != "linux" or python_version < "3.13"
django-cors-headers==4.3.1
gunicorn==21.2.0
whitenoise==6.6.0
pyarrow  # Alternative data processing
